third_party/WebKit/LayoutTests/fast/xmlhttprequest/xmlhttprequest-open-exceptions-expected.txt

CONSOLE ERROR: Refused to connect to 'http://not.example.com/' because it violates the following Content Security Policy directive: "connect-src http://example.com".

This tests that exceptions thrown by XHR.open() have reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhrException.message is "Failed to execute 'open' on 'XMLHttpRequest': 'TRACE' HTTP method is unsupported."
PASS xhrException.message is "Refused to connect to 'http://not.example.com/' because it violates the document's Content Security Policy."
PASS successfullyParsed is true

TEST COMPLETE
